pathways/pyproject.toml
[tool.poetry]
name = "pathways"
version = "0.1.0"
description = ""
authors = ["Iain-S <25081046+Iain-S@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
mdutils = "^1.3.1"
markdown-it-py = "^2.0.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
coverage = "^6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
